files upload

+++ controller/admin.js
@@ -1,7 +1,7 @@
 const Product = require('../models/products');
-const Error = require('../controller/error');
 
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file.js');
 
 const { validationResult } = require('express-validator');
 
@@ -27,13 +27,30 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   console.log(req.body);
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const title = req.body.title;
   const user = req.user;
 
+  console.log('Errrrrrrrrrrrrrrrrrrrr');
+  console.log(image);
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      formsCSS: true,
+      productCSS: true,
+      activeAddProduct: true,
+      editing: false,
+      hasError: true,
+      product: { title: title,  price: price, description: description },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: errors.array()
+    });
+  }
   const errors = validationResult(req);
+  console.log(errors.array());
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
@@ -43,15 +60,17 @@ exports.postAddProduct = (req, res, next) => {
       activeAddProduct: true,
       editing: false,
       hasError: true,
-      product: { title: title, imageUrl: imageUrl, price: price, description: description },
+      product: { title: title,  price: price, description: description },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array()
     });
   }
-  const product = new Product({ title: title, imageUrl: imageUrl, description: description, price: price, userId: user });
+  const imageUrl = image.path;
+  const product = new Product({ title: title, imageUrl:imageUrl, description: description, price: price, userId: user });
   product.save().then(result => {
     res.redirect('/');
   }).catch(err => {
+    console.log('errorrrrr')
     const error = new Error(err);
     error.httpStatusCode = 500;
     next(error); 
@@ -90,21 +109,36 @@ exports.getEditProduct = (req, res, next) => {
 
 exports.postEditProduct = (req, res, next) => {
   const id = req.body.productId;
-  const updatedImageUrl = req.body.imageUrl;
+  const updateImage = req.file;
   const updatedPrice = req.body.price;
   const updatedDescription = req.body.description;
   const updatedTitle = req.body.title;
   const errors = validationResult(req);
-  if (!errors.isEmpty()) {
+
+  if (!updateImage) {
     return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Add Product',
-      path: '/admin/add-product',
+      pageTitle: 'Edit Product',
+      path: '/admin/edit-product',
       formsCSS: true,
       productCSS: true,
       activeAddProduct: true,
       editing: true,
       hasError: true,
-      product: { title: updatedTitle, imageUrl: updatedImageUrl, price: updatedPrice, description: updatedDescription, _id: id },
+      product: { title: title,  price: price, description: description },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: errors.array()
+    });
+  }
+  if (!errors.isEmpty()) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Edit Product',
+      path: '/admin/edit-product',
+      formsCSS: true,
+      productCSS: true,
+      activeAddProduct: true,
+      editing: true,
+      hasError: true,
+      product: { title: updatedTitle, price: updatedPrice, description: updatedDescription, _id: id },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array()
     });
@@ -114,13 +148,16 @@ exports.postEditProduct = (req, res, next) => {
       console.log('not authorized ID');
       return res.redirect('/')
     }
+    if(updateImage){
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = updateImage.path;
+    }
+    deleteFile()
     product.title = updatedTitle;
-    product.imageUrl = updatedImageUrl;
     product.price = updatedPrice;
     product.description = updatedDescription;
     console.log('UPDATED DATA:', product);
     product.save().then(result => {
-
       res.redirect('/');
     });
   }).catch(err => {
@@ -133,13 +170,20 @@ exports.postEditProduct = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const id = req.body.productId;
   console.log('idd', id);
-  Product.deleteOne({ _id: id, userId: req.user._id }).then(result => {
-    res.redirect('/');
-  }).catch(err => {
-    const error = new Error(err);
-    error.httpStatusCode = 500;
-    next(error); 
-  });;
+  Product.findById(id).then(product=>{
+    if(!product){
+      return next(new Error('Product not found'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    Product.deleteOne({ _id: id, userId: req.user._id }).then(result => {
+      res.redirect('/');
+    }).catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      next(error); 
+    });;
+  })
+  
 
 }
